Move some ServiceHelper methods to OperationsHelper
